refactor: move org project-access scoping into SecurityService
The accessibleProjectIds helper lived as a private function on
OrganizationLanguageController, which pulled AuthenticationFacade,
OrganizationRoleService and ProjectService into the controller just to
make an access decision. Move it to SecurityService as
getAccessibleProjectIdsInOrganization so the controller only depends on
SecurityService and the scoping rule is reusable.
