+++ backend/api/src/main/kotlin/io/tolgee/api/v2/controllers/organization/OrganizationLanguageController.kt
@@ -5,11 +5,9 @@ import io.swagger.v3.oas.annotations.tags.Tag
 import io.tolgee.dtos.cacheable.OrganizationLanguageDto
 import io.tolgee.hateoas.language.OrganizationLanguageModel
 import io.tolgee.hateoas.language.OrganizationLanguageModelAssembler
-import io.tolgee.security.authentication.AuthenticationFacade
 import io.tolgee.security.authorization.UseDefaultPermissions
 import io.tolgee.service.language.LanguageService
-import io.tolgee.service.organization.OrganizationRoleService
-import io.tolgee.service.project.ProjectService
+import io.tolgee.service.security.SecurityService
 import org.springdoc.core.annotations.ParameterObject
 import org.springframework.data.domain.Pageable
 import org.springframework.data.domain.Sort
@@ -32,22 +30,8 @@ class OrganizationLanguageController(
   private val languageService: LanguageService,
   private val organizationLanguageModelAssembler: OrganizationLanguageModelAssembler,
   private val pagedOrganizationLanguageAssembler: PagedResourcesAssembler<OrganizationLanguageDto>,
-  private val authenticationFacade: AuthenticationFacade,
-  private val organizationRoleService: OrganizationRoleService,
-  private val projectService: ProjectService,
+  private val securityService: SecurityService,
 ) {
-  private fun accessibleProjectIds(
-    organizationId: Long,
-    requested: List<Long>?,
-  ): List<Long>? {
-    val user = authenticationFacade.authenticatedUser
-    if (organizationRoleService.canUserViewAtLeastMember(user, organizationId)) {
-      return requested
-    }
-    val accessible = projectService.getBelowMemberAccessibleProjectIds(organizationId, user.id).toSet()
-    return requested?.filter { it in accessible } ?: accessible.toList()
-  }
-
   @Operation(
     summary = "Get all languages in use by projects owned by specified organization",
     description = "Returns all languages in use by projects owned by specified organization",
@@ -67,7 +51,7 @@ class OrganizationLanguageController(
     val languages =
       languageService.getPagedByOrganization(
         organizationId,
-        accessibleProjectIds(organizationId, projectIds),
+        securityService.getAccessibleProjectIdsInOrganization(organizationId, projectIds),
         pageable,
         search,
       )
@@ -92,7 +76,7 @@ class OrganizationLanguageController(
     val languages =
       languageService.getBasePagedByOrganization(
         organizationId,
-        accessibleProjectIds(organizationId, projectIds),
+        securityService.getAccessibleProjectIdsInOrganization(organizationId, projectIds),
         pageable,
         search,
       )
